+++ src/Console/ClearCommand.php
@@ -7,7 +7,6 @@ namespace Vellum\Console;
 use Illuminate\Console\Command;
 use Vellum\Cache\CompiledStore;
 use Vellum\Cache\FragmentCache;
-use Vellum\Search\SearchIndexQuery;
 
 /**
  * Deletes the compiled Vellum cache directory.
@@ -23,7 +22,6 @@ final class ClearCommand extends Command
         $store = new CompiledStore((string) config('vellum.cache.path'));
         $store->clear();
         (new FragmentCache)->clear();
-        (new SearchIndexQuery)->clear();
 
         $this->info('Vellum cache cleared.');